resolve some duplicate code issue
