+++ betfund_event_broker/tasks/bet365/__init__.py
@@ -1,5 +1,10 @@
 """Bet365 namespace."""
+from .base import Bet365Task
 from .prematch_odds import Bet365PreMatchOdds
 from .upcoming_events import Bet365UpcomingEvents
 
-__all__ = ["Bet365PreMatchOdds", "Bet365UpcomingEvents"]
+__all__ = [
+    "Bet365Task",
+    "Bet365PreMatchOdds",
+    "Bet365UpcomingEvents"
+]
