+++ betfund_event_broker/tasks/bet365/prematch_odds.py
@@ -1,16 +1,13 @@
 """TODO"""
-import os
+from typing import Union
 
-from betfund_bet365 import Bet365
-from prefect import Task
-from prefect.engine.state import State
+from betfund_bet365.response import Bet365Response
+from betfund_event_broker.tasks.bet365 import Bet365Task
 
 
-class Bet365PreMatchOdds(Task):
+class Bet365PreMatchOdds(Bet365Task):
     """
-    Prefect Task for Bet365 API Request.
-
-    Executes GET request to `PreMatch Odds` endpoint
+    Executes GET request to `PreMatch Odds` endpoint.
 
     Args:
         fi (str): Contains unique identifier for an event
@@ -20,14 +17,9 @@ class Bet365PreMatchOdds(Task):
     Returns:
         State: state of prefect `Task`
     """
+    topic = "preMatchOdds"
 
-    def __init__(self):
-        """Constructor for Bet365UpcomingEvents"""
-        self.api_host = os.getenv("BET365_HOST")
-        self.api_key = os.getenv("BET365_KEY")
-        super().__init__()
-
-    def run(self, fi: str) -> State:
+    def run(self, fi: str) -> Union[None, Bet365Response]:
         """
         Executes API Request to `pre_match_odds(...)` endpoint.
 
@@ -42,22 +34,8 @@ class Bet365PreMatchOdds(Task):
         bet365_client = self._build_client()
 
         if not fi:
-            return None
+            return None, self.topic  # this needs to be amended
 
         response = bet365_client.pre_match_odds(fi=fi)
 
-        print(response)
-
-        return response
-
-    def _build_client(self) -> Bet365:
-        """
-        Build Bet364 API Wrapper Client.
-
-        NOTE: Clients Must be built  outside of `run(...)` for `Task`
-        """
-        bet365_client = Bet365(
-            api_host=self.api_host, api_key=self.api_key
-        )
-
-        return bet365_client
+        return response, self.topic
